requirements.txt
python-binance==1.0.19
typer[all]
click==8.1.7
rich==13.7.1
python-dotenv==1.0.1
